requirements.txt
torch
transformers
ctransformers
gradio
fastapi
python-dotenv
langchain
chromadb
pypdf2
sentence_transformers
tiktoken
tqdm
together
langchain-together
langchain_community
langchain-chroma